requirements.txt
openslide-python>=1.1.2
Pillow>=9.0.0
# tkinter>=8.6
pandas>=1.3.0
numpy>=1.21.0
opencv-python>=4.5.0